[MAINT] Move `Harvester` implementations to dedicated module
